Optimise data memory address de-mapping.
A major bottleneck was found to be the wrapping of *RegionSpec objects in a RegionSpec interface during the Spec method of Region objects. Converting structs to interfaces requires an allocation, a memory-clear and a memory-copy.
Removing this method and moving the code in emulator.go that depended on it into methods on the three region objects has more than halved the total tick time, bringing max emulation speed up to 26 MHz.